requirements.txt
asgiref==3.6.0
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.1.7
django-crispy-forms==2.0
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.12.0
Flask==2.3.2
h11==0.14.0
idna==3.4
instaloader==4.9.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.0
openpyxl==3.1.2
outcome==1.2.0
pandas==2.0.3
Pillow==9.4.0
platformdirs==3.3.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2022.7
urllib3==2.0.3
virtualenv==20.22.0
virtualenv2==0.0.1
virtualenvwrapper-win==1.2.7
Werkzeug==2.3.6
wsproto==1.2.0
